Refactor : Moved fn calls to helper fns
